Making `derive_transcription_src_path` generally available
