+++ emitters/yaml_class_emitter.cpp
@@ -71,60 +71,6 @@ bool yaml_class_emitter::do_merge(const std::string& filepath,
 
 /**************************************************************************************************/
 
-std::filesystem::path derive_transcription_src_path(const std::filesystem::path& dst,
-                                                    const std::string& title) {
-    const auto parent = dst.parent_path();
-    std::size_t best_match = std::numeric_limits<std::size_t>::max();
-    std::filesystem::path result;
-    const std::string& current_version = hyde_version();
-
-    for (const auto& entry : std::filesystem::directory_iterator(dst.parent_path())) {
-        const auto sibling = entry.path();
-        if (!is_directory(sibling)) continue;
-        const auto index_path = sibling / index_filename_k;
-
-        if (!exists(index_path)) {
-            std::cerr << "WARN: expected " << index_path.string() << " but did not find one\n";
-            continue;
-        }
-
-        const auto have_docs = parse_documentation(index_path, true);
-
-        if (have_docs._error) {
-            std::cerr << "WARN: expected " << index_path.string() << " to have docs\n";
-            continue;
-        }
-
-        const auto& have = have_docs._json;
-
-        if (have.count("version")) {
-            // Transcription is when we're going from a previous version of hyde to this one.
-            // So if the versions match, this is a directory that has already been transcribed.
-            // (Transcribing from a newer version of hyde docs to older ones isn't supported.)
-            if (static_cast<const std::string&>(have["version"]) == current_version) {
-                continue;
-            }
-        }
-
-        // REVISIT (fosterbrereton): Are these titles editable? Would
-        // users muck with them and thus break this algorithm?
-        const std::string& have_title = static_cast<const std::string&>(have["title"]);
-
-        // score going from what we have to what this version computed.
-        const auto match = diff_score(have_title, title);
-        if (match > best_match) {
-            continue;
-        }
-
-        best_match = match;
-        result = sibling;
-    }
-
-    return result;
-}
-
-/**************************************************************************************************/
-
 bool yaml_class_emitter::emit(const json& j, json& out_emitted, const json& inherited) {
     json node = base_emitter_node("class", j["name"], "class", has_json_flag(j, "implicit"));
     node["hyde"]["defined_in_file"] = defined_in_file(j["defined_in_file"], _src_root);
@@ -168,11 +114,7 @@ bool yaml_class_emitter::emit(const json& j, json& out_emitted, const json& inhe
         // we are now trying to load and reconcile with what we've created. In this case, we can
         // assume the "shape" of the documentation is the same, which means that within the parent
         // folder of `dst` is the actual source folder that holds the old documentation, just under
-        // a different name. So, iterate the list of directories in the parent, find their
-        // `index.md` files, load the `title` of each, and find the best-fit against the `title`
-        // stored in `node`. Once we have found that best fit directory, rename it to the `dst`
-        // path, and then we can continue. It's a lot of legwork, but the only way I can think of
-        // to reconcile a previous version name for a symbol with a new one.
+        // a different name. Find that folder and rename it.
 
         std::filesystem::rename(derive_transcription_src_path(dst, node["title"]), dst);
     }
@@ -180,15 +122,18 @@ bool yaml_class_emitter::emit(const json& j, json& out_emitted, const json& inhe
     bool failure =
         reconcile(std::move(node), _dst_root, std::move(dst) / index_filename_k, out_emitted);
 
-    const auto& methods = j["methods"];
     yaml_function_emitter function_emitter(_src_root, _dst_root, _mode, _options, true);
+    auto emitted_methods = hyde::json::array();
+    const auto& methods = j["methods"];
 
     for (auto it = methods.begin(); it != methods.end(); ++it) {
         function_emitter.set_key(it.key());
         auto function_emitted = hyde::json::object();
         failure |= function_emitter.emit(it.value(), function_emitted, out_emitted.at("hyde"));
-        out_emitted["methods"].push_back(std::move(function_emitted));
+        emitted_methods.push_back(std::move(function_emitted));
     }
+
+    out_emitted["methods"] = std::move(emitted_methods);
 
     return failure;
 }
